# server/src/swagger/routes.yaml
swagger: '3.0.0'
info:
  title: Kids Activity Tracker API
  version: '1.0.0'
  description: |
    Comprehensive API documentation for the Kids Activity Tracker application.
    
    ## API Overview
    
    ### Core Concepts:
    - **Categories**: 5 age-based groupings (Early Years: Parent Participation, Early Years: On My Own, School Age, Youth, All Ages & Family)
    - **Activity Types**: 22 subject-based groupings (Swimming & Aquatics, Team Sports, etc.)
    - **Global Filters**: hideClosedActivities, hideFullActivities (apply to all endpoints except Favorites)
    - **Pagination**: 50 items per page with offset-based pagination
    
    ### Navigation Flows:
    1. **Browse by Location**: Dashboard → Cities → Venues → Activities
    2. **Browse by Category**: Dashboard → Categories → Category Detail → Activities  
    3. **Browse by Activity Type**: Dashboard → Activity Types → Activity Type Detail → Activities
    4. **Recommendations**: Dashboard → Recommended Activities (user-specific)
    
    ### React Native Components Usage:
    - **DashboardScreen**: Uses categories, activity types, locations APIs
    - **CategoryDetailScreen**: Uses `/categories/{id}/activities`
    - **AllCategoriesScreen**: Uses `/categories`
    - **LocationBrowseScreen**: Uses `/locations/cities`, `/locations/{cityId}/venues`
    - **ActivityDetailScreen**: Uses `/activities/{id}`, favorites APIs
    - **SearchScreen**: Uses `/activities/search` with extensive filtering
paths:
...
  /api/v1/activities:
    get:
      tags: [Activities]
      summary: Search and filter activities
      description: |
        Primary activity search endpoint with comprehensive filtering.
        Used by: SearchScreen, FilterScreen, RecommendedActivitiesScreen
        Supports all global filters and advanced search capabilities.
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: Search in activity name, description
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of activities to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of activities to skip
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
          description: Filter out closed activities
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
          description: Filter out full activities
        - in: query
          name: activityTypes
          schema:
            type: array
            items:
              type: string
          description: Filter by activity type IDs or names
        - in: query
          name: categories
          schema:
            type: array
            items:
              type: string
          description: Filter by category IDs
        - in: query
          name: locations
          schema:
            type: array
            items:
              type: string
          description: Filter by location IDs or city names
        - in: query
          name: ageMin
          schema:
            type: integer
          description: Minimum age filter
        - in: query
          name: ageMax
          schema:
            type: integer
          description: Maximum age filter
        - in: query
          name: costMin
          schema:
            type: number
          description: Minimum cost filter
        - in: query
          name: costMax
          schema:
            type: number
          description: Maximum cost filter
      responses:
        200:
          description: Paginated search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
                  filters:
                    type: object
                    description: Applied filters summary
  /api/v1/activities/{id}:
    get:
      tags: [Activities]
      summary: Get activity details
      description: |
        Get detailed information about a specific activity.
        Used by: ActivityDetailScreen, ActivityCard navigation
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Activity UUID
      responses:
        200:
          description: Activity details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  activity:
                    $ref: '#/components/schemas/Activity'
        404:
          description: Activity not found
...
  /api/v1/activity-types/{id}/activities:
    get:
      tags: [Activity Types]
      summary: Get activities for specific activity type
      description: |
        Returns paginated activities that belong to a specific activity type.
        Used by: ActivityTypeDetailScreen
        Navigation flow: Dashboard → Activity Types → Activity Type Detail (this endpoint)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Activity Type UUID
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: Paginated activities for activity type
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  activityType:
                    $ref: '#/components/schemas/ActivityType'
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
...
  /api/v1/activity-types/{typeCode}/subtypes/{subtypeCode}/activities:
    get:
      tags: [Activity Types]
      summary: Get activities for specific activity subtype
      description: |
        Get paginated activities for a specific activity subtype.
        Used by: ActivitySubtypeDetailScreen
      parameters:
        - in: path
          name: typeCode
          required: true
          schema:
            type: string
          description: Activity Type code
        - in: path
          name: subtypeCode
          required: true
          schema:
            type: string
          description: Activity Subtype code
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: Activities for subtype
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  activitySubtype:
                    type: object
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        404:
          description: Activity subtype not found

  # ==================== CHILDREN ENDPOINTS ====================
...
  /api/child-activities/{childId}/activities/{activityId}:
    put:
      tags: [Child Activities]
      summary: Update activity status for child
      description: |
        Update the status of an activity link for a child.
        Used by: Activity management screens
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: activityId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [interested, registered, completed, cancelled]
                notes:
                  type: string
      responses:
        200:
          description: Activity link updated
        404:
          description: Link not found
    
    delete:
      tags: [Child Activities]
      summary: Remove activity link from child
      description: |
        Remove the link between a child and an activity.
        Used by: Activity management screens
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
        - in: path
          name: activityId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Link removed successfully
        404:
          description: Link not found
...
  /api/child-activities/{childId}/activities:
    get:
      tags: [Child Activities]
      summary: Get activities for a specific child
      description: |
        Get all activities associated with a specific child.
        Used by: Child activity screens
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: status
          schema:
            type: string
            enum: [interested, registered, completed, cancelled]
          description: Filter by activity status
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: Child's activities
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  child:
                    $ref: '#/components/schemas/Child'
                  activities:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        404:
          description: Child not found

  # ==================== LOCATIONS ENDPOINTS ====================
...
  /api/v1/locations/{locationId}/activities:
    get:
      tags: [Locations]
      summary: Get activities at a specific venue/location
      description: |
        Returns paginated activities at a specific venue/location.
        Used by: VenueDetailScreen or LocationDetailScreen.
        Navigation flow: Cities → Venues → Activities (this endpoint) → Activity Detail
      parameters:
        - in: path
          name: locationId
          required: true
          schema:
            type: string
            format: uuid
          description: Location UUID
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
            maximum: 100
          description: Number of activities to return
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
          description: Number of activities to skip for pagination
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [dateStart, name, cost, registrationDate]
            default: dateStart
          description: Field to sort by
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: asc
          description: Sort direction
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
          description: Filter out closed activities
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
          description: Filter out full activities
        - in: query
          name: search
          schema:
            type: string
          description: Search in activity name and description
        - in: query
          name: ageMin
          schema:
            type: integer
          description: Minimum age requirement
        - in: query
          name: ageMax
          schema:
            type: integer
          description: Maximum age requirement
        - in: query
          name: costMin
          schema:
            type: number
          description: Minimum cost filter
        - in: query
          name: costMax
          schema:
            type: number
          description: Maximum cost filter
        - in: query
          name: startDate
          schema:
            type: string
            format: date
          description: Filter activities starting after this date
        - in: query
          name: endDate
          schema:
            type: string
            format: date
          description: Filter activities ending before this date
        - in: query
          name: activityTypes
          schema:
            type: array
            items:
              type: string
          description: Filter by activity type IDs or names
        - in: query
          name: categories
          schema:
            type: array
            items:
              type: string
          description: Filter by category IDs
      responses:
        200:
          description: Paginated activities at location
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  location:
                    $ref: '#/components/schemas/Location'
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        404:
          description: Location not found

  # ==================== CATEGORIES ENDPOINTS ====================
...
  /api/v1/categories/{id}/activities:
    get:
      tags: [Categories]
      summary: Get activities for specific age-based category
      description: |
        Returns paginated activities that belong to a specific age-based category.
        Used by: CategoryDetailScreen for displaying activities in a category.
        Navigation flow: Dashboard → Categories → Category Detail (this endpoint) → Activity Detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
          description: Category UUID
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [dateStart, name, cost, registrationDate]
            default: dateStart
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: asc
        - in: query
          name: hideClosedActivities
          schema:
            type: boolean
            default: false
        - in: query
          name: hideFullActivities
          schema:
            type: boolean
            default: false
        - in: query
          name: search
          schema:
            type: string
          description: Search in activity name and description
        - in: query
          name: ageMin
          schema:
            type: integer
        - in: query
          name: ageMax
          schema:
            type: integer
        - in: query
          name: costMin
          schema:
            type: number
        - in: query
          name: costMax
          schema:
            type: number
        - in: query
          name: locations
          schema:
            type: array
            items:
              type: string
          description: Filter by location IDs or names
      responses:
        200:
          description: Paginated activities for category
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  category:
                    $ref: '#/components/schemas/Category'
                  activities:
                    type: array
                    items:
                      $ref: '#/components/schemas/Activity'
                  pagination:
                    $ref: '#/components/schemas/PaginationResponse'
        404:
          description: Category not found

  # ==================== INVITATIONS ENDPOINTS ====================
...
  /api/shared-activities/children/{childId}/activities:
    get:
      tags: [Shared Activities]
      summary: Get activities for specific shared child
      description: |
        Get activities for a specific shared child.
        Used by: Shared child activity screens
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: childId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        200:
          description: Shared child activities
...
  /api/v1/cities/{city}/activities:
    get:
      tags: [Cities]
      summary: Get activities in city
      description: |
        Get all activities in a specific city with pagination.
        Used by: City activities screens
      parameters:
        - in: path
          name: city
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        200:
          description: City activities

  # ==================== REFERENCE ENDPOINTS ====================
...
components:
  schemas:
    Activity:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        cost:
          type: number
        spotsAvailable:
          type: integer
        totalSpots:
          type: integer
        registrationStatus:
          type: string
          enum: [Open, Closed, Full, Cancelled]
        dateStart:
          type: string
          format: date-time
        dateEnd:
          type: string
          format: date-time
        registrationDate:
          type: string
          format: date-time
        location:
          $ref: '#/components/schemas/Location'
        activityType:
          $ref: '#/components/schemas/ActivityType'
        provider:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            website:
              type: string
    ActivityType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        activityCount:
          type: integer
        subtypes:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
              code:
                type: string
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          enum: [
            "Early Years: Parent Participation",
            "Early Years: On My Own", 
            "School Age",
            "Youth",
            "All Ages & Family"
          ]
        description:
          type: string
        ageMin:
          type: integer
        ageMax:
          type: integer
        requiresParent:
          type: boolean
        activityCount:
          type: integer
    Location:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        address:
          type: string
        city:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            province:
              type: string
            country:
              type: string
        activityCount:
          type: integer
...
    Child:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        dateOfBirth:
          type: string
          format: date
        age:
          type: integer
          description: Calculated age in years
        interests:
          type: array
          items:
            type: string
        notes:
          type: string
        isActive:
          type: boolean
...
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
          description: Total number of items
        limit:
          type: integer
          description: Items per page
        offset:
          type: integer
          description: Items skipped
        pages:
          type: integer
          description: Total number of pages
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...